Add files via upload
Project Objectives
• Understand the key drivers for why some people are more expensive (i.e., require more 
health care), as well as predict which people will be expensive (in terms of health care 
costs).
• Predict people who will spend a lot of money on health care next year (i.e., which people 
will have high healthcare costs).
• Provide actionable insight to the HMO, in terms of how to lower their total healthcare 
costs, by providing a specific recommendation on how to lower healthcare costs.
Project Agenda
To finish this project, we have five phases. 
1) Phase one is to clean the original data since we find out that there are some not available 
values in the dataset. We want to make sure that the results of the project are more 
accurate after cleaning.
2) Phase two is exploring the data by histograms and boxplots of numeric. 
3) Phase three is an overview of the important variables. In this data, we have 12 
independent variables in total, but we want to know which variables are affecting the 
total cost more or the most. Finding out the important variables can help us to make the 
suggestions effectively.
4)  Phase four is doing the data modelling and trends for predicting if a person is “expensive”. 
The last phase is giving an accurate conclusion and recommendations.